Adds a PullRequestEventManager class.
The PullRequestEventManager is responsible for handling
the payload of pull request webhook sent by Github. It
dispatch the payload to a class that knows how to handle
the specific action of the pull request event.

Possible actions are "assigned", "unassigned", "labeled",
"unlabeled", "opened", "closed".

The only supported action (for now) is opened and the class
that handles that payload is
PullRequestEventManager::OpenedPullRequest. It creates a new
pull request object. A spec is also added for this class.

Github payload for testing are loading using JSON fixtures.
A module has been configured in RSpec to make JSON fixtures
loading easier.